refactor: modifies the ResponseEntity of controller layer and modifies the method 'deleteTitle'

+++ src/main/java/Book/demo/Books/BooksService.java
@@ -10,10 +10,7 @@ import org.springframework.web.server.ResponseStatusException;
 import org.yaml.snakeyaml.events.Event;
 
 import javax.imageio.plugins.tiff.ExifInteroperabilityTagSet;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 import java.util.stream.Collector;
 import java.util.stream.Collectors;
 
@@ -53,6 +50,11 @@ public class BooksService {
     }
 
     public void deleteTitle(Long id){
+
+        if (booksRepository.findById(id).isEmpty()){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+
         booksRepository.deleteById(id);
     }
